backend/requirements.txt
asgiref==3.10.0
behave==1.3.3
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==7.0.0
Django==5.2.7
filelock==3.20.0
fsspec==2025.9.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
parse==1.20.2
parse_type==0.6.6
psutil==7.1.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.1
djangorestframework-simplejwt==5.3.1
whitenoise==6.4.0

